a07a314e: are timer counts updated?
After adding a new timer, the counts on the home page should be
update to reflect this.

I spent a bit of time here in the test code to make it easier
to do similar tests in future.

Some points to note:

* I started a very large timer so that it should still be active
  even when we finish testing.

* Scraping HTML like this isn't very clean, but we'll accept it here

* Note the use of is_deeply for comparing two structures

* Also, the test line is much more readable than using content_like,
  and we can pick out the expected counter values quickly.

I'm also continuing to use "magic numbers". I generate these by calling
uuid and copying the first 8 characters. Their main use is just to have
a consistent "name" for a test case that I can search for in the test
script if something goes wrong, but sometimes I will also embed them in
the code (as a cross-reference to the test case being worked on).

Sometimes I will also have my code emit a magic number to satisfy a
test. Usually in the case where we want to show that there was an
error or that something worked, but we don't want to test for a
specific string. I'll talk about these a bit more later.