Add boilerplate code for using a Mojo::IOLoop timer and update 'started' count.
If you run the tests again, you'll see that our new test isn't passing.
I'm checking the code in anyway so that we can examine what's going on.

After the test failing the first time, I could see that the number
of 'started' timers was still stuck on 0 the next time I loaded the main
page. To check to make sure that my code was actually updating this
total, I added a debug message to the /add_timer routine and checked:

    $ ./eventy get /add_timer/99999
    [2019-08-10 14:04:04.47029] [9509] [debug] GET "/add_timer/99999" (b96c0a44)
    [2019-08-10 14:04:04.47078] [9509] [debug] Routing to a callback
    [2019-08-10 14:04:04.47107] [9509] [debug] 200 OK (0.000778s, 1285.347/s)
    Added a timer of 99999 seconds.
    Set global 'started' count to 1

So it seems that the new code *is* actually updating the global variable
correctly, but between the call to /add_timer and reloading /, this change
is being forgotten. What's going on?

Intuitively, it seems that what's happening is that each time we call
$t->get() in our test script, we're starting off again with a completely
fresh server state. So changes *do* seem to made, but they aren't persisting.

As I said, I'm just pausing here before figuring out the problem. Can
you figure out what the solution is?