A surprising result.
I had assumed that the way I was creating the initial test object
was causing it to create a clean server state for each call. With
that in mind, I tried running it as a daemon by hand (both in
the test script and outside). It turns out that this didn't solve
the problem.

What *did* solve the problem was embedding the data that we want
to become stateful into the application itself.

I also added a line in the timer callback that appears in the
server log if you're running the server as a daemon. It makes it
clearer what the boilerplate code actually does.