Write code to satisfy test. (+Recap)
Note that this code is slightly unsafe because the callback is
holding onto a reference to $c->app->{timer_stats}. If that
value changes before the callback gets run, the callback will
be using the *old* value. It's not a problem here, but you should
be aware that what I'm doing here (reusing $stats) in two contexts
isn't good style and can lead to memory leaks and/or buggy behaviour.

To recap, we've seen how to:

* use timers, the simplest kind of asynchronous event programming
* stash global data that we want to persist server side!
* track totals of started, active and elapsed timers

There are a couple of directions we can head in from here.

Firstly, we can do a bit of code refactoring. If we want to create
other timers in other parts of our code and make sure that the
server is keeping its timer counts up to date, then it's inconvenient
to have to duplicate the counter-related code everywhere.

Secondly, we can expand the program to use more features that are
available from Mojo::IOLoop. In particular, in order to understand
event-driven programming in Mojo, you have to get to grips with
Emitters and Reactors.

I'll park the question of which path we'll take for now. The reason
being that I have a clear idea of what the *final* program will look
like. Eventually, the program will satisfy both of the above. It will
use more advanced Mojo::IOLoop features and it will (hopefully) have
a clean, nicely-encapsulated design.

While we're resting here at a crossroad, I'd just like to make a brief
note about how I'm applying test-driven design, and how it will inform
the next steps that I take.

One important concept of TDD that applies here is that as you add
more tests, all of the existing tests should continue to pass. There's
also another principle that I'm trying to stick to, which is that
every new test that you add should fail first. You write the test case
first, and *then* add the functionality. TDD also allows for refactoring
after you've managed to get some functionality to work. However, refactoring
that actually *adds* new functionality is strictly forbidden.

As a result, there's sometimes a delicate balancing act required
to plot a course from where you currently are with a program to
getting to where you want to be, while still keeping with the principles
of test-driven design.

While I ponder that next step, I can at least give you some
idea of where I want to get to. When the program is finished:

* parts of the program that use timers can do so without worrying
  about updating counts

* we'll have a new app-wide timer creation factory, but its
  syntax and semantics will be exactly like the built-in factory

* we'll separate out the logic of creating and tracking timers
  and timer events into two distinct parts that will communicate
  using Mojo events:

    * encapsulation of a "noisy" timer factory that emits events
      when timers start or elapse

    * code in our main program that subscribes to these events
      and keeps track of counts.