For a bit of background, look at the man page for Mojo::EventEmitter. This code is based on example found at the top of the page.
What's happening here is:

* We define a new class to emit events
* We subscribe our main code to the "started" event
* When main code consumes the event, it updates some counters
* Where the timer was actually created, we use a method call
  on the Trampoline object, which triggers the started event

We've gone some way towards encapsulating the counting behaviour,
moving (some of it) it closer to the main package.

We've also got a clear separation between the event emitter
which is just emitting events that might be of interest, and
the main program that uses those events in a (somewhat) meaningful
way.