Finished writing conformance tests.
From here on, everything is mostly refactoring.

However, there is one change that I'd like to make, which is to
make all the event stuff completely transparent to the users of
timers.

The code to do this will add new functionality, but it won't change
the external interface that we already have. That makes it a little
bit difficult to write a test in such a way that it will fail the
first time we call it, but there is a way. Basically, we can
introspect into the running server so we can check if it has the
particular internal interface that we're implementing. In fact, we
can go further and actually eval code within the server.

I'll get to that in a moment, but the next change will be to rename
our Trampoline class.